utils/requirements.txt
streamlit
openai
PyPDF2
huggingface_hub
pinecone-client
langchain